+++ src/Troschuetz.Random/TRandom.cs
@@ -1746,8 +1746,8 @@ namespace Troschuetz.Random
         ///   Returns an unsigned random number.
         /// </summary>
         /// <returns>
-        ///   A 32-bit unsigned integer greater than or equal to <see cref="uint.MinValue"/> and less
-        ///   than or equal to <see cref="uint.MaxValue"/>.
+        ///   A 32-bit unsigned integer greater than or equal to 0, and less than
+        ///   <see cref="uint.MaxValue"/>; that is, the range of return values includes 0 but not <see cref="uint.MaxValue"/>.
         /// </returns>
         public uint NextUInt() => Generator.NextUInt();
 
@@ -1758,7 +1758,7 @@ namespace Troschuetz.Random
         ///   A 32-bit unsigned integer greater than or equal to 0, and less than
         ///   <see cref="uint.MaxValue"/>; that is, the range of return values includes 0 but not <see cref="uint.MaxValue"/>.
         /// </returns>
-        public uint NextUIntExclusiveMaxValue() => Generator.NextUIntExclusiveMaxValue();
+        public uint NextUIntExclusiveMaxValue() => Generator.NextUInt();
 
         /// <summary>
         ///   Returns an unsigned random number.